refactor(engine): move the builtin engine plugins under src/engine
The whatsapp-web-js and baileys engine plugins are engine-internal
adapters wired in by EngineFactory, not runtime-installable plugins —
nothing reads the source tree's plugins/ directory from disk. Colocate
them with the engine they plug into as engine/builtin/, keeping their
manifests and colocated specs; the now-empty src/plugins/ tree is gone
and the one stale path comment is updated.
