+++ src/engine/builtin/baileys/manifest.json
