+++ src/engine/builtin/whatsapp-web-js/manifest.json
