+++ src/engine/engine.factory.ts
@@ -5,8 +5,8 @@ import { ConfigService } from '@nestjs/config';
 import { IWhatsAppEngine } from './interfaces/whatsapp-engine.interface';
 import { WhatsAppWebJsAdapter } from './adapters/whatsapp-web-js.adapter';
 import { PluginLoaderService, PluginType, IEnginePlugin, PluginManifest } from '../core/plugins';
-import { WhatsAppWebJsPlugin } from '../plugins/engines/whatsapp-web-js';
-import { BaileysPlugin } from '../plugins/engines/baileys';
+import { WhatsAppWebJsPlugin } from './builtin/whatsapp-web-js';
+import { BaileysPlugin } from './builtin/baileys';
 import { createLogger } from '../common/services/logger.service';
 import { BaileysMessageStoreService } from './adapters/baileys-message-store.service';
 import { LidMappingStoreService } from './identity/lid-mapping-store.service';
